// src/app/models/user.ts
export interface Profile {
  name: string;
  age: number;
  work: string;
  hobbies: string[];
  featured_phrases: string[];
}

// {
//   "nombre": "Juan",
//   "edad": 30,
//   "ocupación": "programador",
//   "intereses": ["fútbol", "tecnología", "viajes"],
//   "frases_destacadas": [
//     "Ayer estuve trabajando en un proyecto de Angular.",
//     "El fútbol es mi pasatiempo favorito, siempre apoyo a mi equipo local."
//   ]
// }